+++ src/services/index.ts
@@ -5,3 +5,4 @@ export * from "./TileWebGLService";
 export * from "./ContourWebGLService";
 export * from "./ScriptingService";
 export * from "./CatalogWebGLService";
+export * from "./CatalogApiService";
